main
Make sure to install the necessary dependencies (requests, beautifulsoup4, pandas) before running the script. Additionally, you may need to customize the code further based on the structure of the e-commerce site you are scraping.

Note that continuous scraping of a website may be against the website's terms of service or may put a strain on their servers. Ensure that you are complying with the website's policies and use this script responsibly.